[flutter_tools] remove globals from Device and separate FlutterDeviceManager
