+++ packages/flutter_tools/lib/src/fuchsia/fuchsia_device.dart
@@ -64,6 +64,7 @@ Future<void> _kDefaultDartDevelopmentServiceStarter(
     0,
     true,
     disableServiceAuthCodes,
+    logger: globals.logger,
   );
 }
