+++ packages/flutter_tools/lib/src/resident_runner.dart
@@ -271,6 +271,7 @@ class FlutterDevice {
             ddsPort,
             ipv6,
             disableServiceAuthCodes,
+            logger: globals.logger,
           );
         } on dds.DartDevelopmentServiceException catch (e, st) {
           if (!allowExistingDdsInstance ||
