+++ packages/flutter_tools/test/commands.shard/hermetic/attach_test.dart
@@ -94,7 +94,7 @@ void main() {
           .thenAnswer((_) async {});
         when(device.dds).thenReturn(mockDds);
         final Completer<void> noopCompleter = Completer<void>();
-        when(mockDds.startDartDevelopmentService(any, any, false, any)).thenReturn(null);
+        when(mockDds.startDartDevelopmentService(any, any, false, any, logger: anyNamed('logger'))).thenReturn(null);
         when(mockDds.uri).thenReturn(Uri.parse('http://localhost:8181'));
         when(mockDds.done).thenAnswer((_) => noopCompleter.future);
 
@@ -323,7 +323,7 @@ void main() {
       )).thenReturn(mockHotRunner);
       when(mockHotRunner.exited).thenReturn(false);
       when(mockHotRunner.isWaitingForObservatory).thenReturn(false);
-      when(mockDds.startDartDevelopmentService(any, any, false, any)).thenReturn(null);
+      when(mockDds.startDartDevelopmentService(any, any, false, any, logger: anyNamed('logger'))).thenReturn(null);
       when(mockDds.uri).thenReturn(Uri.parse('http://localhost:8181'));
 
       testDeviceManager.addDevice(device);
@@ -410,7 +410,7 @@ void main() {
       )).thenReturn(mockHotRunner);
       when(mockHotRunner.exited).thenReturn(false);
       when(mockHotRunner.isWaitingForObservatory).thenReturn(false);
-      when(mockDds.startDartDevelopmentService(any, any, false, any)).thenReturn(null);
+      when(mockDds.startDartDevelopmentService(any, any, false, any, logger: anyNamed('logger'))).thenReturn(null);
       when(mockDds.uri).thenReturn(Uri.parse('http://localhost:8181'));
 
       testDeviceManager.addDevice(device);
@@ -457,7 +457,7 @@ void main() {
           .thenAnswer((_) async {});
         when(device.dds)
           .thenReturn(mockDds);
-        when(mockDds.startDartDevelopmentService(any, any, any, any))
+        when(mockDds.startDartDevelopmentService(any, any, any, any, logger: anyNamed('logger')))
           .thenReturn(null);
         when(mockDds.uri).thenReturn(Uri.parse('http://localhost:8181'));
         final Completer<void> noopCompleter = Completer<void>();
